requirements.txt
requests>=2.26.0
beautifulsoup4~=4.10.0
pytz>=2021.3
icalendar>=4.0.9
python-dateutil~=2.8.1
openpyxl~=3.0.5
